FloodNav-Path-Planning/requirements_fastapi.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.7.2
Pillow==10.0.1